+++ src/API/index.js
@@ -1,0 +1,2 @@
+export * as Mappings from './Mappings';
+export * as Configurations from './Configurations';