+++ src/Configurations/ListPane/index.js
@@ -1,0 +1,1 @@
+export { ListPane } from './ListPane';